fixed bug of receiving scores of old generation

+++ plugin/argos_loop_functions/mpga_phototaxis_obstacle_loop_functions.h
@@ -36,7 +36,7 @@ public:
 
 private:
 
-    int m_iScore;
+    float m_iScore;
 };
 
 #endif
